# src/main/resources/Openapi/CreditModule-openapi_loan.yaml
openapi: "3.1.0"
info:
  title: "CreditModule API"
  description: "CreditModule API"
  version: "1.0.0"
servers:
  - url: "https://CreditModule"
paths:
  /loans:
    post:
      summary: "POST loans"
      operationId: "createLoan"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Loan"
    get:
      summary: "GET loans"
      operationId: "listLoansByCustomer"
      parameters:
        - name: "customerId"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Loan"
  /loans/{loanId}/installments:
    get:
      summary: "GET loans/{loanId}/installments"
      operationId: "listInstallments"
      parameters:
        - name: "loanId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/LoanInstallment"
  /loans/{loanId}/pay:
    post:
      summary: "POST loans/{loanId}/pay"
      operationId: "payLoan"
      parameters:
        - name: "loanId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "amount"
          in: "query"
          required: true
          schema:
            $ref: "#/components/schemas/BigDecimal"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
components:
  schemas:
    BigDecimal:
      type: "object"
      properties: { }
    CreateLoanRequest:
      type: "object"
      properties:
        customerId:
          type: "integer"
          format: "int64"
        loanAmount:
          $ref: "#/components/schemas/BigDecimal"
        interestRate:
          $ref: "#/components/schemas/BigDecimal"
        numberOfInstallments:
          type: "integer"
          format: "int32"
    Customer:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        surname:
          type: "string"
        creditLimit:
          $ref: "#/components/schemas/BigDecimal"
        usedCreditLimit:
          $ref: "#/components/schemas/BigDecimal"
    Loan:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        customer:
          $ref: "#/components/schemas/Customer"
        loanAmount:
          $ref: "#/components/schemas/BigDecimal"
        numberOfInstallment:
          type: "integer"
          format: "int32"
        createDate:
          type: "string"
          format: "date-time"
    LoanInstallment:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        loan:
          $ref: "#/components/schemas/Loan"
        amount:
          $ref: "#/components/schemas/BigDecimal"
        paidAmount:
          $ref: "#/components/schemas/BigDecimal"
        dueDate:
          type: "string"
          format: "date"
        paymentDate:
          type: "string"
          format: "date"
    PaymentResponse:
      type: "object"
      properties:
        installmentsPaid:
          type: "integer"
          format: "int32"
        totalPaid:
          $ref: "#/components/schemas/BigDecimal"